requirements.txt
pyrogram==1.3.6
TgCrypto
python-dotenv
motor
pillow
aiofiles
aiohttp
requests
pymongo[srv]
psutil
emoji
tglogging
